refactor: rename post_process files for clarity (#224)
- post_process.ts → post_process_cli.ts (CLI entry point)
- post_processing.ts → comment_builder.ts (GitHub comment builders)
- Update all imports, build config, action.yml paths, and README
